Refactor import statements to use relative paths for consistency
