+++ pages_bots/ncc_page.py
@@ -36,7 +36,7 @@ home_dir = os.getenv("HOME")
 from ..super.S_API import bot_api
 from ..super.S_Category import catdepth_new
 from ..super.S_Page import super_page
-from .api_utils.user_agent import default_user_agent
+from ..api_utils.user_agent import default_user_agent
 from ..api_utils import lang_codes
 
 from ..accounts.user_account_ncc import User_tables, SITECODE, FAMILY
