+++ pages_bots/page.py
@@ -32,8 +32,8 @@ from ..super.S_API import bot_api
 from ..super.S_Page import super_page
 from ..super.S_Category import catdepth_new
 from ..accounts import useraccount
-from .api_utils.user_agent import default_user_agent
-from .api_utils import lang_codes
+from ..api_utils.user_agent import default_user_agent
+from ..api_utils import lang_codes
 
 home_dir = os.getenv("HOME")
 tool = home_dir.split("/")[-1] if home_dir else None
